requirements.txt
click >=7
tomli
uvicorn
fastapi
requests
pyaml
nextflow